+++ packages/lambda.0.1.7/src/parse.ml
@@ -1,0 +1,7 @@
+exception InvalidLambdaString;;
+
+let parse s = 
+  try 
+    Parser.prog Lexer.lexer (Lexing.from_string s)
+  with | _ -> raise InvalidLambdaString
+;;