+++ packages/lambda.0.1.7/src/parse.mli
@@ -1,0 +1,3 @@
+exception InvalidLambdaString
+val parse: string -> L.term
+(** [parse s] converts s to a term, or raise InvalidLambdaString on failures *)