+++ packages/ocaml-version.2.5.0/CHANGES.md
@@ -1,3 +1,9 @@
+## v2.5.0 (2020-03-22)
+
+* Add OCaml 4.09.1 and 4.10.0 releases (@avsm @kit-ty-kate).
+* Add `equal` function to test versions for equality (#7 @kit-ty-kate)
+* Add a top-level printer hint for `Ocaml_version.t` (@avsm)
+
 ## v2.4.0 (2020-02-20)
 
 * Add `unreleased_betas` list to enumerate the latest release-candidate